Cleaned up the logic around classloading
